refactor(types): rename AlertLevel.ts
