+++ src/components/auth/forgot-password-common/index.tsx
@@ -7,8 +7,7 @@ import { VERIFICATION_CODE_LENGTH, isValidUsername } from '@/utils/auth';
 
 import AuthForm from '../auth-form';
 
-import { AuthFormFieldEnum } from '@/types';
-import AlertLevel from '@/types/alert-level-enum';
+import { AuthFormFieldEnum, AlertLevel } from '@/types';
 import AuthType from '@/types/auth-type-enum';
 
 type ForgotPasswordCommonProps = {
