+++ src/components/auth/sign-up-common/index.tsx
@@ -7,8 +7,7 @@ import { authService } from '@/services';
 
 import AuthForm from '../auth-form';
 
-import { AuthFormFieldEnum } from '@/types';
-import AlertLevel from '@/types/alert-level-enum';
+import { AuthFormFieldEnum, AlertLevel } from '@/types';
 
 export default function SignUpCommon() {
   const {
