+++ src/types/AlertLevel.ts
@@ -1,0 +1,6 @@
+export enum AlertLevel {
+  INFO,
+  WARNING,
+  ERROR,
+  SUCCESS,
+}
