--- src/types/alert-level-enum.ts
@@ -1,8 +1,0 @@
-enum AlertLevel {
-  INFO,
-  WARNING,
-  ERROR,
-  SUCCESS,
-}
-
-export default AlertLevel;
